requirements.txt
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
#colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.1.4
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.0
#decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
easyocr==1.7.2
#exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing==2.0.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
fuzzywuzzy==0.18.0
idna==3.10
imageio==2.36.0
#importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
#packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.3
parso==0.8.3
pickleshare==0.7.5
pillow==11.0.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
#psutil @ file:///D:/bld/psutil_1729847175714/work
pure-eval==0.2.2
pyclipper==1.3.0.post6
Pygments==2.16.1
pyparsing==3.2.0
pytesseract==0.3.13
python-bidi==0.6.3
#python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
#six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
tornado==6.3.3
traitlets==5.11.2
#typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.8
#zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
